+++ CHANGELOG.md
@@ -68,7 +68,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- **Icon** `SimpleBasket`.
+- **Icon** `Trophy`,`SimpleBasket`.
 
 ### Fixed